refactored arg parsing
